feat(hastag): improve hashtag (#1486)
